basic_block coverage: gets stuck in infinite loop
The binary search can get stuck in an infinite loop when `mid` is
assigned to `lo` and the first two checks on the tb/bb coverage do not
pass. In this case, `lo` will always be the same value as `mid`, which can
never be greater than `hi` and therefore the `while` loop will never
exit.